removed imguizmo sequencer files
